\name{testFuns}
\alias{testEnsembl}
\cr
\alias{testUcsc}
\alias{testUcscAll}
\cr
\alias{testUcscUtr}
\alias{testUcscUtrAll}
\cr
\alias{testCustomGtf}
\cr
\alias{testKnownBuild}
\alias{testCustomBuild}
\title{Query and database build testing functions}
\usage{
    testEnsembl(level = c("normal", "long", "short"),
        versioned = FALSE)
    \cr
    testUcsc(orgs, refdbs, types, versioned = FALSE)
    testUcscAll()
    \cr
    testUcscUtr(orgs, refdbs, versioned = FALSE)
    testUcscUtrAll()
    \cr
    testCustomGtf(gtf)
    \cr
    testKnownBuild(org, refdb, ver = NULL)
    testCustomBuild(gtf, metadata)
    
}
\arguments{
    \item{level}{}
    
    \item{orgs}{}
    
    \item{refdbs}{}
    
    \item{types}{}
    
    \item{org}{}
    
    \item{refdb}{}
    
    \item{gtf}{}
    
    \item{metadata}{}
    
    \item{versioned}{}
}
\value{
    This function 
}
\description{
    
}
\examples{

}
\author{
    Panagiotis Moulos
}

